chore: adds /delete router for deleting address book
